:sparkles: add automated linting and formatting functions using ruff
