+++ scripts/utils/__init__.py
@@ -1,3 +1,4 @@
 from .pr import create_update_pr
+from .automated_ruff import lint_path, format_path
 
-__all__ = ("create_update_pr",)
+__all__ = ("create_update_pr", "lint_path", "format_path")
